Add exception to author Real Name policy
This matches what was decided at the Eclipse Architecture Council
in March 2020 which was minuted as follows:
A GitHub account as a contributor is ok, it can be traced back to an individual.
An ECA must be signed in any case. This requires a real email address and this is sufficient.
EMO expectation to committers is to monitor and catch/report shenanigans.
The handbook wording needs an updated and will be investigated separately.